pyproject.toml
[tool.poetry]
name = "whisper-asr-webservice"
version = "1.2.0"
description = "Whisper ASR Webservice is a general-purpose speech recognition webservice."
homepage  = "https://github.com/cbsiamlg/whisper-asr-webservice/"
authors = [
    "Justin Nguyen",
    "Zachariah Mustafa",
    "Adam Leary",
]
readme = "README.md"
packages = [{ include = "app" }]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.10"
unidecode = "^1.3.4"
uvicorn = { extras = ["standard"], version = "^0.18.2" }
gunicorn = "^20.1.0"
tqdm = "^4.64.1"
transformers = "^4.22.1"
python-multipart = "^0.0.5"
ffmpeg-python = "^0.2.0"
fastapi = "^0.95.1"
llvmlite = "^0.40.1"
numba = "^0.57.1"
openai-whisper = "20231117"
faster-whisper = "^1.0.0"
torch = { version = "^1.13.1", extras = ["cpu"] }

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.poetry.group.test.dependencies]
pytest = "^6.2.5"
pytest-mock = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
